growthMonitoring_ml/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
pillow>=10.0
numpy>=1.24
opencv-python>=4.8
# Model runtime
torch
torchvision
